+++ tests/utsuho/test_half_to_full.py
@@ -27,7 +27,30 @@ from utsuho import ConverterConfig, HalfToFullConverter
     ('\uFF9E', '゛'), ('\uFF9F', '゜'),
     ('ｶﾞﾞ', 'ガ゛'), ('ﾞﾞ', '゛゛'), ('ﾑﾞ', 'ム゛'),
     ('ﾊﾟﾟ', 'パ゜'), ('ﾟﾟ', '゜゜'), ('ﾑﾟ', 'ム゜'),
-    ('｢AｱｶﾞB漢｣', '「AアガB漢」'), ('「ＡアガＢ漢」', '「ＡアガＢ漢」'),
+    ('｢AｱｶﾞB漢｣', '「ＡアガＢ漢」'), ('「ＡアガＢ漢」', '「ＡアガＢ漢」'),
+    ('0', '０'), ('1', '１'), ('2', '２'), ('3', '３'), ('4', '４'),
+    ('5', '５'), ('6', '６'), ('7', '７'), ('8', '８'), ('9', '９'),
+    ('0\uFE00', '０\uFE00'), ('\uFE001\uFE00', '\uFE00１'),
+    ('A', 'Ａ'), ('B', 'Ｂ'), ('C', 'Ｃ'), ('D', 'Ｄ'), ('E', 'Ｅ'),
+    ('F', 'Ｆ'), ('G', 'Ｇ'), ('H', 'Ｈ'), ('I', 'Ｉ'), ('J', 'Ｊ'),
+    ('K', 'Ｋ'), ('L', 'Ｌ'), ('M', 'Ｍ'), ('N', 'Ｎ'), ('O', 'Ｏ'),
+    ('P', 'Ｐ'), ('Q', 'Ｑ'), ('R', 'Ｒ'), ('S', 'Ｓ'), ('T', 'Ｔ'),
+    ('U', 'Ｕ'), ('V', 'Ｖ'), ('W', 'Ｗ'), ('X', 'Ｘ'), ('Y', 'Ｙ'),
+    ('Z', 'Ｚ'),
+    ('a', 'ａ'), ('b', 'ｂ'), ('c', 'ｃ'), ('d', 'ｄ'), ('e', 'ｅ'),
+    ('f', 'ｆ'), ('g', 'ｇ'), ('h', 'ｈ'), ('i', 'ｉ'), ('j', 'ｊ'),
+    ('k', 'ｋ'), ('l', 'ｌ'), ('m', 'ｍ'), ('n', 'ｎ'), ('o', 'ｏ'),
+    ('p', 'ｐ'), ('q', 'ｑ'), ('r', 'ｒ'), ('s', 'ｓ'), ('t', 'ｔ'),
+    ('u', 'ｕ'), ('v', 'ｖ'), ('w', 'ｗ'), ('x', 'ｘ'), ('y', 'ｙ'),
+    ('z', 'ｚ'),
+    ('!', '！'), ('"', '＂'), ('#', '＃'), ('$', '＄'), ('%', '％'),
+    ('&', '＆'), ('\'', '＇'), ('(', '（'), (')', '）'), ('*', '＊'),
+    ('+', '＋'), (',', '，'), ('-', '－'), ('.', '．'), ('/', '／'),
+    (':', '：'), (';', '；'), ('<', '＜'), ('=', '＝'), ('>', '＞'),
+    ('?', '？'), ('@', '＠'), ('[', '［'), ('\\', '＼'), (']', '］'),
+    ('^', '＾'), ('_', '＿'), ('`', '｀'), ('{', '｛'), ('|', '｜'),
+    ('}', '｝'), ('~', '～'),
+    ('\u0020', '　'), ('\u00A0', '　'),
 ])
 def test_half_to_full(s, expect):
     cnv = HalfToFullConverter()
@@ -50,7 +73,7 @@ def test_half_to_full(s, expect):
     if s == '｢' or s == '｣':
         assert actual == s
     elif s == '｢AｱｶﾞB漢｣':
-        assert actual == '｢AアガB漢｣'
+        assert actual == '｢ＡアガＢ漢｣'
     else:
         assert actual == expect
 
@@ -69,6 +92,47 @@ def test_half_to_full(s, expect):
 
     if s == 'ｰ':
         assert actual == s
+    else:
+        assert actual == expect
+
+    config = ConverterConfig(space=False)
+    cnv = HalfToFullConverter(config)
+    actual = cnv.convert(s)
+
+    if s in ['\u0020', '\u00A0'] or s == '〜':
+        assert actual == s
+    else:
+        assert actual == expect
+
+    config = ConverterConfig(ascii_symbol=False)
+    cnv = HalfToFullConverter(config)
+    actual = cnv.convert(s)
+
+    if s in [chr(c) for c in range(ord('!'), ord('/') + 1)] \
+            or s in [chr(c) for c in range(ord(':'), ord('@') + 1)] \
+            or s in [chr(c) for c in range(ord('['), ord('`') + 1)] \
+            or s in [chr(c) for c in range(ord('{'), ord('~') + 1)]:
+        assert actual == s
+    else:
+        assert actual == expect
+
+    config = ConverterConfig(ascii_digit=False)
+    cnv = HalfToFullConverter(config)
+    actual = cnv.convert(s)
+
+    if s in [chr(c) for c in range(ord('0'), ord('9') + 1)] or s == '0\uFE00' or s == '\uFE001\uFE00':
+        assert actual == s
+    else:
+        assert actual == expect
+
+    config = ConverterConfig(ascii_alphabet=False)
+    cnv = HalfToFullConverter(config)
+    actual = cnv.convert(s)
+
+    if s in [chr(c) for c in range(ord('A'), ord('Z') + 1)] or s in [chr(c) for c in range(ord('a'), ord('z') + 1)]:
+        assert actual == s
+    elif s == '｢AｱｶﾞB漢｣':
+        assert actual == '「AアガB漢」'
     else:
         assert actual == expect