+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Utsuho Changes
 
+## Version 1.1.0
+
+Released on 2023-04-05.
+
+* Added the feature to convert between half-width spaces and full-width spaces.
+* Added the feature to convert between half-width symbols and full-width symbols.
+* Added the feature to convert between half-width digits and full-width digits.
+* Added the feature to convert between half-width alphabets and full-width alphabets.
+
 ## Version 1.0.0
 
 Released on 2023-04-02.